requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
psycopg2-binary==2.9.9
click==8.3.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
loguru==0.7.3
mako==1.3.10
markupsafe==3.0.3
motor==3.7.1
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.11.0
pymongo==4.15.3
python-dotenv==1.2.1
python-jose==3.5.0
redis==7.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.49.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
python-multipart==0.0.20